# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KnowledgeGPT API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/documents/upload-url": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Create Upload Url",
        "description": "Step 1 of upload: register a PENDING document row and return a\npresigned S3 PUT URL. The client uploads the file bytes directly to S3,\nthen calls POST /confirm.",
        "operationId": "create_upload_url_api_documents_upload_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadUrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/documents/confirm": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Confirm Upload",
        "description": "Step 2 of upload: client confirms the S3 PUT finished. We verify the\nobject actually exists (never trust the client alone) and enqueue\nbackground processing.",
        "operationId": "confirm_upload_api_documents_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/documents/{document_id}/status": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document Status",
        "description": "Polled by the frontend upload progress UI.",
        "operationId": "get_document_status_api_documents__document_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_api_documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knowledge Base Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentPublic"
                  },
                  "title": "Response List Documents Api Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}": {
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Delete Document",
        "operationId": "delete_document_api_documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmUploadRequest": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "ConfirmUploadRequest"
      },
...
      "DocumentPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "knowledge_base_id": {
            "type": "string",
            "format": "uuid",
            "title": "Knowledge Base Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "file_type": {
            "$ref": "#/components/schemas/FileType"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "status_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Detail"
          },
          "processing_progress_pct": {
            "type": "integer",
            "title": "Processing Progress Pct"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "original_size_bytes": {
            "type": "integer",
            "title": "Original Size Bytes"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "knowledge_base_id",
          "name",
          "file_type",
          "status",
          "processing_progress_pct",
          "chunk_count",
          "original_size_bytes",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentPublic"
      },
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "virus_scanning",
          "extracting",
          "ocr_processing",
          "chunking",
          "embedding",
          "completed",
          "failed"
        ],
        "title": "DocumentStatus"
      },
      "DocumentStatusResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "status_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Detail"
          },
          "processing_progress_pct": {
            "type": "integer",
            "title": "Processing Progress Pct"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "processing_progress_pct"
        ],
        "title": "DocumentStatusResponse"
      },
      "FileType": {
        "type": "string",
        "enum": [
          "pdf",
          "docx",
          "doc",
          "txt",
          "csv",
          "xlsx",
          "xls",
          "pptx",
          "markdown",
          "html",
          "xml",
          "json",
          "rtf",
          "image",
          "zip"
        ],
        "title": "FileType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadUrlRequest": {
        "properties": {
          "knowledge_base_id": {
            "type": "string",
            "format": "uuid",
            "title": "Knowledge Base Id"
          },
          "filename": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "description": "Client-reported size; re-validated server-side after upload."
          }
        },
        "type": "object",
        "required": [
          "knowledge_base_id",
          "filename",
          "content_type",
          "size_bytes"
        ],
        "title": "UploadUrlRequest"
      },
      "UploadUrlResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "s3_key": {
            "type": "string",
            "title": "S3 Key"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "upload_url",
          "s3_key",
          "expires_in"
        ],
        "title": "UploadUrlResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}